Finally all working.
After finding that the issue was caused by trying to read
in PORT.. instead of PIN.., I re-enabled the code to check
on all INTx lines (not only D0), just to be on the safe side.

We switch off just fast enough to not mess with other readings
after Y8. 1 cycle more and we would be stepping on other reads.

There is Y9 and Y10, and althought they are not used, stepping
on those with PORTC from another column causes characters on the
screen. So it's important to switch off fast enough.